interface/requirements.txt
Pillow
grpcio==1.39.0
grpcio-tools==1.39.0
numpy
